react refactor: Move components from App.js to components packages
