+++ src/components/CatForm.js
@@ -1,0 +1,39 @@
+import React from "react";
+
+const includesHangul = (text) => /[ㄱ-ㅎ|ㅏ-ㅣ|가-힣]/i.test(text);
+
+function CatForm(props) {
+    const [value, setValue] = React.useState('')
+    const [errorMessage, setErrorMessage] = React.useState('')
+
+    function handleInputChange(e) {
+        const userValue = e.target.value;
+        setErrorMessage('')
+        if (includesHangul(userValue)) {
+            setErrorMessage('No Korean')
+        }
+        setValue(userValue.toUpperCase())
+    }
+
+    function handleFormSubmit(e) {
+        e.preventDefault();
+
+        if (value === '') {
+            setErrorMessage('No Blank')
+            return
+        }
+        props.updateMainCat(value)
+        props.initHeart()
+    }
+
+    return (
+        <form onSubmit={handleFormSubmit}>
+            <input type="text" name="name" placeholder="영어 대사를 입력해주세요" value={value}
+                   onChange={handleInputChange}/>
+            <button type="submit">생성</button>
+            <p style={{color: 'red'}}>{errorMessage}</p>
+        </form>
+    )
+}
+
+export default CatForm;
