+++ src/components/Favorites.js
@@ -1,0 +1,24 @@
+const uuidv4 = () => {
+    return ([1e7] + -1e3 + -4e3 + -8e3 + -1e11).replace(/[018]/g, c =>
+        (c ^ crypto.getRandomValues(new Uint8Array(1))[0] & 15 >> c / 4).toString(16)
+    );
+}
+
+function CatItem(props) {
+    return (
+        <li>
+            <img src={props.img}/>
+        </li>
+    )
+}
+
+function Favorites(props) {
+
+    return (
+        <ul className="favorites">
+            {props.favorites.map(cat => <CatItem key={uuidv4()} img={cat}/>)}
+        </ul>
+    )
+}
+
+export default Favorites;
