+++ src/components/MainCard.js
@@ -1,0 +1,11 @@
+const MainCard = (props) => {
+    return (
+        <div className="main-card">
+            <img src={props.img} alt={props.alt}/>
+            <button onClick={props.onHeartClick}>{props.heart}
+            </button>
+        </div>
+    )
+}
+
+export default MainCard;
