+++ src/components/Title.js
@@ -1,0 +1,7 @@
+const Title = (props) => {
+    return (
+        <h1>{props.children}</h1>
+    )
+}
+
+export default Title;
